requirements.txt
streamlit
pathway
google-genai
python-dotenv
spacy

